[Tech] refacto table component
