+++ frontend/src/components/Table/index.tsx
@@ -1,0 +1,120 @@
+import { ChevronIcon } from '@features/commonStyles/icons/ChevronIcon.style'
+import { Icon, SimpleTable } from '@mtes-mct/monitor-ui'
+import { flexRender, getCoreRowModel, getSortedRowModel, useReactTable, type SortingState } from '@tanstack/react-table'
+import { useVirtualizer } from '@tanstack/react-virtual'
+import { forwardRef, useCallback, useState } from 'react'
+import styled from 'styled-components'
+
+export function TableWithRef({ columns, data }, ref) {
+  const { current: currentRef } = ref
+
+  const [sorting, setSorting] = useState<SortingState>([{ desc: true, id: 'updatedAt' }])
+
+  const table = useReactTable({
+    columns,
+    data,
+    enableSortingRemoval: false,
+    getCoreRowModel: getCoreRowModel(),
+    getSortedRowModel: getSortedRowModel(),
+    onSortingChange: setSorting,
+    state: {
+      sorting
+    }
+  })
+
+  const { rows } = table.getRowModel()
+  const rowVirtualizer = useVirtualizer({
+    count: rows.length,
+    estimateSize: () => 40,
+    getItemKey: useCallback((index: number) => `${rows[index]?.id}`, [rows]),
+    getScrollElement: () => currentRef,
+    overscan: 10
+  })
+
+  const virtualRows = rowVirtualizer.getVirtualItems()
+  const [paddingTop, paddingBottom] =
+    virtualRows.length > 0
+      ? [
+          Math.max(0, virtualRows[0]?.start ?? 0),
+          Math.max(0, rowVirtualizer.getTotalSize() - (virtualRows[virtualRows.length - 1]?.end ?? 0))
+        ]
+      : [0, 0]
+
+  return (
+    <StyledDasboardsContainer ref={ref}>
+      <SimpleTable.Table>
+        <SimpleTable.Head>
+          {table.getHeaderGroups().map(headerGroup => (
+            <tr key={headerGroup.id}>
+              {headerGroup.headers.map(header => (
+                <SimpleTable.Th key={header.id} $width={header.column.getSize()}>
+                  {header.isPlaceholder ? undefined : (
+                    <SimpleTable.SortContainer
+                      className={header.column.getCanSort() ? 'cursor-pointer select-none' : ''}
+                      onClick={header.column.getToggleSortingHandler()}
+                    >
+                      {flexRender(header.column.columnDef.header, header.getContext())}
+
+                      {header.column.getCanSort() &&
+                        ({
+                          asc: <StyledChevronIcon $isOpen={false} $right={false} />,
+                          desc: <StyledChevronIcon $isOpen $right={false} />
+                        }[header.column.getIsSorted() as string] ?? <Icon.SortingArrows size={14} />)}
+                    </SimpleTable.SortContainer>
+                  )}
+                </SimpleTable.Th>
+              ))}
+            </tr>
+          ))}
+        </SimpleTable.Head>
+        <tbody>
+          {paddingTop > 0 && (
+            <tr>
+              <td aria-label="empty-line-for-scroll" style={{ height: `${paddingTop}px` }} />
+            </tr>
+          )}
+          {virtualRows.map(virtualRow => {
+            const row = rows[virtualRow.index]
+
+            return (
+              <SimpleTable.BodyTr key={virtualRow.key}>
+                {row?.getVisibleCells().map(cell => (
+                  <SimpleTable.Td
+                    key={cell.id}
+                    $isCenter={cell.column.id === 'geom' || cell.column.id === 'edit'}
+                    style={{
+                      maxWidth: cell.column.getSize(),
+                      minWidth: cell.column.getSize(),
+                      padding: cell.column.id === 'geom' || cell.column.id === 'edit' ? '0px' : '10px 12px',
+                      width: cell.column.getSize()
+                    }}
+                  >
+                    {flexRender(cell.column.columnDef.cell, cell.getContext())}
+                  </SimpleTable.Td>
+                ))}
+              </SimpleTable.BodyTr>
+            )
+          })}
+          {paddingBottom > 0 && (
+            <tr>
+              <td aria-label="empty-line-for-scroll" style={{ height: `${paddingBottom}px` }} />
+            </tr>
+          )}
+        </tbody>
+      </SimpleTable.Table>
+    </StyledDasboardsContainer>
+  )
+}
+
+export const Table = forwardRef(TableWithRef)
+
+const StyledDasboardsContainer = styled.div`
+  overflow: auto;
+  width: fit-content;
+  // scroll width (~15px) + 4px
+  padding-right: 19px;
+`
+const StyledChevronIcon = styled(ChevronIcon)`
+  margin-top: 0px;
+  margin-right: 0px;
+`
